Review PipedCommand and CommandList
Previously we wanted to send a list of commands as CommandList
and command with pipe as PipedCommand.
After more discussion, we will create one shell script that contains
list of command and return the path of shell script.

Due to this decision, let's remove PipedCommand and CommandList.

ONE-vscode-DCO-1.0-Signed-off-by: Hyun Sik Yoon <hyunsik.yoon.1024@gmail.com>