[CI] Make CI fail when license verification is failed
Until now, CI did not fail even when license verification is failed.
It was because maintainers can judge the result by comment.
However, as result of license verification is not created anymore,
it is better for CI to fail when the verification is failed.

ONE-vscode-DCO-1.0-Signed-off-by: Seok NamKoong <seok9311@naver.com>